Stub out Buildkite pipeline
Consider this provisional.

It uses a Buildkite account as well as a Quay.io account that no one
else has access to. However, if merged, the following image will be
publicly-accessible and should be kept up to date on merges to main.

    quay.io/colemanx/raku-doc-website:latest

I'd like to ask around the other community projects and see if there's
anything else to integrate with. In the meantime, this moves the ball
forward.