Add EBook compilation to website build + container
The build script drops the ebook file in our repository root. We can
move it into the rendered_html folder before archiving. This way, when
the archive is unpacked inside the container, it's located at the
expected root url.

The buildscript should be adjusted to put the ebook in rendered_html
at some point. But this hack works for now.